updated the ui to support the quaternion rotation overhaul
UI rotation of the scene camera stays the same because even though the camera can now freely rotate anywhere, we don't want it to. But entity rotation now rotates relatively based on the amount of change which allows free rotation anywhere without gimbal lock. The displayed euler values in the UI are no longer 1:1 and can output strange values sometimes.